requirements.txt
aiosqlite==0.19.0
fastapi==0.100.1
databases==0.7.0
SQLAlchemy==1.4.49
pytest-cov==4.1.0
uvicorn==0.23.1
httpx==0.24.1